Improve changeFont function and add new function
i improve the changeFont function to decrement and increment the
font size if it's too big or small. i also make new function calculate
for efficiency so i don't have to write the same code multiple times.
I also add a function that will separate the operand if the first
operand length reached the max limit. another simple addition is a new
parameter that will automatically output the second operand on the next
line when the firstOperand is too long.